Move ResultSummaryReporter from model to GUI presenters namespace
